Add ITermConceptClientMappingSvc interface to represent terminology translate functions
